Changes to merge SQLite and Redis storage writer #3024
