+++ plaso/storage/redis/writer.py
@@ -2,11 +2,11 @@
 """Storage writer for Redis."""
 
 from plaso.lib import definitions
-from plaso.storage import interface
+from plaso.storage import writer
 from plaso.storage.redis import redis_store
 
 
-class RedisStorageWriter(interface.StorageWriter):
+class RedisStorageWriter(writer.StorageWriter):
   """Redis-based storage writer."""
 
   def __init__(
@@ -29,27 +29,6 @@ class RedisStorageWriter(interface.StorageWriter):
 
     self._redis_namespace = '{0:s}-{1:s}'.format(
         task.session_identifier, task.identifier)
-    self._store = None
-
-  def AddAttributeContainer(self, container):
-    """Adds an attribute container.
-
-    Args:
-      container (AttributeContainer): attribute container.
-    """
-    self._store.AddAttributeContainer(container)
-
-    if container.CONTAINER_TYPE == self._CONTAINER_TYPE_ANALYSIS_REPORT:
-      self._UpdateAnalysisReportSessionCounter(container)
-
-    elif container.CONTAINER_TYPE == self._CONTAINER_TYPE_ANALYSIS_WARNING:
-      self.number_of_analysis_warnings += 1
-
-    elif container.CONTAINER_TYPE == self._CONTAINER_TYPE_EVENT:
-      self._UpdateEventParsersSessionCounter(container)
-
-    elif container.CONTAINER_TYPE == self._CONTAINER_TYPE_EVENT_DATA:
-      self._UpdateEventDataParsersMappings(container)
 
   def Close(self):
     """Closes the storage writer.
@@ -63,30 +42,6 @@ class RedisStorageWriter(interface.StorageWriter):
 
     self._store.Finalize()
     self._store = None
-
-  def GetAttributeContainerByIdentifier(self, container_type, identifier):
-    """Retrieves a specific type of container with a specific identifier.
-
-    Args:
-      container_type (str): container type.
-      identifier (AttributeContainerIdentifier): attribute container identifier.
-
-    Returns:
-      AttributeContainer: attribute container or None if not available.
-    """
-    return self._store.GetAttributeContainerByIdentifier(
-        container_type, identifier)
-
-  def GetAttributeContainers(self, container_type):
-    """Retrieves a specific type of attribute containers.
-
-    Args:
-      container_type (str): attribute container type.
-
-    Returns:
-      generator(AttributeContainers): attribute container generator.
-    """
-    return self._store.GetAttributeContainers(container_type)
 
   # pylint: disable=redundant-returns-doc,useless-return
   def GetFirstWrittenEventSource(self):
@@ -123,28 +78,6 @@ class RedisStorageWriter(interface.StorageWriter):
 
     return None
 
-  def GetSortedEvents(self, time_range=None):
-    """Retrieves the events in increasing chronological order.
-
-    This includes all events written to the storage including those pending
-    being flushed (written) to the storage.
-
-    Args:
-      time_range (Optional[TimeRange]): time range used to filter events
-          that fall in a specific period.
-
-    Returns:
-      generator(EventObject): event generator.
-
-    Raises:
-      IOError: if the storage writer is closed.
-      OSError: if the storage writer is closed.
-    """
-    if not self._store:
-      raise IOError('Unable to read from closed storage writer.')
-
-    return self._store.GetSortedEvents(time_range=time_range)
-
   # pylint: disable=arguments-differ
   def Open(self, redis_client=None, **unused_kwargs):
     """Opens the storage writer.
@@ -160,6 +93,12 @@ class RedisStorageWriter(interface.StorageWriter):
         storage_type=self._storage_type,
         session_identifier=self._task.session_identifier,
         task_identifier=self._task.identifier)
+
+    if self._serializers_profiler:
+      self._store.SetSerializersProfiler(self._serializers_profiler)
+
+    if self._storage_profiler:
+      self._store.SetStorageProfiler(self._storage_profiler)
 
     self._store.Open(redis_client=redis_client)
 
@@ -180,22 +119,6 @@ class RedisStorageWriter(interface.StorageWriter):
           information.
     """
     raise IOError('Preprocessing information not supported by the redis store.')
-
-  def SetSerializersProfiler(self, serializers_profiler):
-    """Sets the serializers profiler.
-
-    Args:
-      serializers_profiler (SerializersProfiler): serializers profiler.
-    """
-    self._serializers_profiler = serializers_profiler
-
-  def SetStorageProfiler(self, storage_profiler):
-    """Sets the storage profiler.
-
-    Args:
-      storage_profiler (StorageProfiler): storage profiler.
-    """
-    self._storage_profiler = storage_profiler
 
   def WritePreprocessingInformation(self, knowledge_base):
     """Writes preprocessing information.
@@ -247,37 +170,3 @@ class RedisStorageWriter(interface.StorageWriter):
           start.
     """
     raise IOError('Session start is not supported by the redis store.')
-
-  def WriteTaskCompletion(self, aborted=False):
-    """Writes task completion information.
-
-    Args:
-      aborted (Optional[bool]): True if the session was aborted.
-
-    Raises:
-      IOError: if the storage type is not supported or
-          when the storage writer is closed.
-      OSError: if the storage type is not supported or
-          when the storage writer is closed.
-    """
-    if self._storage_type != definitions.STORAGE_TYPE_TASK:
-      raise IOError('Only task storage is supported by the redis store.')
-
-    self._task.aborted = aborted
-    task_completion = self._task.CreateTaskCompletion()
-    self._store.WriteTaskCompletion(task_completion)
-
-  def WriteTaskStart(self):
-    """Writes task start information.
-
-    Raises:
-      IOError: if the storage type does not support writing a task
-          start or when the storage writer is closed.
-      OSError: if the storage type does not support writing a task
-          start or when the storage writer is closed.
-    """
-    if self._storage_type != definitions.STORAGE_TYPE_TASK:
-      raise IOError('Only task storage is supported by the redis store.')
-
-    task_start = self._task.CreateTaskStart()
-    self._store.WriteTaskStart(task_start)
